requirements.txt
python-telegram-bot==22.5
sqlalchemy==2.0.45
aiosqlite==0.22.1
python-dotenv==1.2.1
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21
opencv-python==4.12.0.88
tensorflow==2.20.0
numpy==2.2.6